Fix trait’s actions
